cover

+++ tests/test_nd_viewer.py
@@ -48,7 +48,7 @@ def test_ndviewer(qtbot: QtBot, backend: str, monkeypatch: pytest.MonkeyPatch) -
 # not testing pygfx yet...
 @pytest.mark.skipif(sys.platform != "darwin", reason="the mouse event is tricky")
 def test_hover_info(qtbot: QtBot) -> None:
-    data = np.ones((4, 3, 32, 32), dtype=np.uint8)
+    data = np.ones((4, 3, 32, 32), dtype=np.float32)
     viewer = NDViewer(data)
     qtbot.addWidget(viewer)
     viewer.show()
